+++ agents/sirji_agents/llm/orchestrator.py
@@ -2,11 +2,11 @@ import textwrap
 import os
 import json
 
-# TODO - log file should be dynamically created based on agent ID
 from sirji_tools.logger import o_logger as logger
 
-from sirji_messages import message_parse, MessageParsingError, MessageValidationError, ActionEnum, AgentEnum, allowed_response_templates, permissions_dict
+from sirji_messages import message_parse, ActionEnum, AgentEnum, allowed_response_templates, permissions_dict, MessageIncorrectFormatError, MessageMultipleActionError, MessageUnRecognizedActionError, MessageMissingPropertyError, MessageLengthConstraintError
 from .model_providers.factory import LLMProviderFactory
+from ..decorators import retry_on_exception
 
 class Orchestrator():
     def __init__(self, agent_output_folder_index):
@@ -56,15 +56,27 @@ class Orchestrator():
                 parsed_response_message = message_parse(response_message)
                 conversation.append({"role": "assistant", "content": response_message, "parsed_content": parsed_response_message})
                 break
-            except (MessageParsingError, MessageValidationError) as e:
-            # Handling both MessageParsingError and MessageValidationError similarly
-                logger.info("Error while parsing the message.\n")
-                retry_llm_count += 1
-                if retry_llm_count > 2:
-                    raise e
-                logger.info(f"Requesting LLM to resend the message in correct format.\n")
-                conversation.append({"role": "assistant", "content": response_message, "parsed_content": {}})
-                conversation.append({"role": "user", "content": "Error obtained in processing your last response. Your response must conform strictly to one of the allowed Response Templates, as it will be processed programmatically and only these templates are recognized. Your response must be enclosed within '***' at the beginning and end, without any additional text above or below these markers. Not conforming above rules will lead to response processing errors."})
+            except (MessageIncorrectFormatError, MessageMultipleActionError, MessageUnRecognizedActionError, MessageMissingPropertyError, MessageLengthConstraintError ) as e:
+                if isinstance(e, MessageIncorrectFormatError):
+                    logger.info("Error while parsing the message.\n")
+                    retry_llm_count += 1
+                    if retry_llm_count > 2:
+                        raise e
+                    logger.info(f"Requesting LLM to resend the message in correct format.\n")
+                    conversation.append({"role": "assistant", "content": response_message, "parsed_content": {}})
+                    conversation.append({"role": "user", "content": "Error obtained in processing your last response. Your response must conform strictly to one of the allowed Response Templates, as it will be processed programmatically and only these templates are recognized. Your response must be enclosed within '***' at the beginning and end, without any additional text above or below these markers. Not conforming above rules will lead to response processing errors."})
+                if isinstance(e, MessageMissingPropertyError):
+                    conversation.append({"role": "assistant", "content": response_message, "parsed_content": {}})
+                    conversation.append({"role": "user", "content": "Error obtained in processing your last response. your response is missing some required properties. Please refer to the allowed response templates and provide the missing properties."})
+                if isinstance(e, MessageUnRecognizedActionError):
+                    conversation.append({"role": "assistant", "content": response_message, "parsed_content": {}})
+                    conversation.append({"role": "user", "content": "Error obtained in processing your last response. Your response contains an unrecognized action. Please refer to the allowed response templates and provide a valid action."})
+                if isinstance(e, MessageLengthConstraintError):
+                    conversation.append({"role": "assistant", "content": response_message, "parsed_content": {}})
+                    conversation.append({"role": "user", "content": "Error obtained in processing your last response. Your response does not have all the required properties."})
+                if isinstance(e, MessageMultipleActionError):
+                    conversation.append({"role": "assistant", "content": response_message, "parsed_content": {}})
+                    conversation.append({"role": "user", "content": "Error obtained in processing your last response. Your response contains more than one ACTION keyword."})
             except Exception as e:
                 logger.info(f"Generic error while parsing message. Error: {e}\n")
                 raise e
@@ -72,6 +84,7 @@ class Orchestrator():
             
         return response_message, prompt_tokens, completion_tokens
     
+    @retry_on_exception(logger=logger)
     def __call_llm(self, conversation):
         history = []
 
@@ -118,7 +131,7 @@ class Orchestrator():
         
         pseudo_code = textwrap.dedent(f"""
             Pseudo code which you must follow:
-                1. Invoke agent RECIPE_SELECTOR to get the recipe selected from the available recipes by SIRJI_USER and then store it in Agent Output Folder.
+                1. Invoke agent {ActionEnum.RECIPE_SELECTOR.name} to get the recipe selected from the available recipes by SIRJI_USER and then store it in Agent Output Folder.
                 2. Read the selected recipe from the Agent Output Folder.
                 3. Loop over the prescribed tasks in the selected recipe and invoke the agent specified in the recipe alongside the current task, explaining the task in the BODY of the invocation.
             """)